[WIP][FEATURE] Update Backend URLs when navigating
Wouldn't it be cool if editors could share their current URL just by clicking in the browser?

This patch adds a iframe listener so every time a module gets changed, the URL in the browser
is updated.

We use history.replaceState to decorate iframe state
changes with an according URL in the browser URL bar.
We avoid history.pushState as that would result in
two state changes for every new iframe load.

Note: Browser back button does not fire popstate when
navigating between iframes, therefore we handle
module navigation as result of the iframe load.

Every iframe is expected to indicate it's module via the data-module-name
tag. Therefore all core modules are adapted to fill the module name
(which was already available since ever, but mostly unused – only
EXT:form used it for customized styling, this is changed to a css class)

To Be Done:
+ push/replaceState should remove the token query parameter
+ push/replaceState also should work for non-modules / regular routes
+ resolve active module menu entry when using browser back button
- evaluate the query parameter (re-auth) when the link is used in another browser
- Adding tests when a user visits a non-authenticated URL, authenticate (and show error or redirect etc)
- Fix popups (in EditDocumentController)
- Test .htaccess check when TYPO3 is installed in subfolders
- Test Install Tool
- Fix module name advertisement in install tool backend module controller

Not needed right now, because browser handles history for module iframes automatically
- Resolve URLs handled in push state
- Resolve URLs when using the browser back button

Resolves: #?
Releases: master
Change-Id: I52b4d97de1e488b4978a95a618753d3bbbbda881