[WIP][FEATURE] Include lit-html as client-side template engine
Adds the lit-html template engine, that originated from
the polymer project and is now base for lit-element
and many web components.

ContextMenu and LiveSearch are adapted as first examples to
show lit-html rendering.
ContextMenu is an example for rendering directly to an
HTML element. LiveSearch renders to string.

TODO:
 * Feature RST

Commands:
  yarn add lit-html
  yarn add --dev grunt-rollup
  grunt build

Resolves: #
Related: #91810
Releases: master
Change-Id: If419d378d6e98c150b219fe441b0766d12e9b639