[FEATURE] Introduce lit-html and lit-element as client-side template engine
Add the lit-html template engine, that originated from
the polymer project and is now base for web components
together with the web component base case LitElement.

It is used to replace string-concatenation/jquery based
templating in order to adapt to modern web standards.

Notification, ContextMenu and LiveSearch are adapted as
first examples to show lit-html and lit-element rendering.
Notification is an example for a web component via LitElement.
ContextMenu is now rendered directly to the
HTML element, LiveSearch renders to string.

typo3/testing-framework is adapted to include lit-html and
lit-element with:
https://github.com/TYPO3/testing-framework/pull/221
(releases as 6.6.0 and merged in #93074)

Commands:

  yarn add lit-html@^1.3 lit-element@^2.4
  yarn add --dev grunt-rollup
  grunt build

Resolves: #91810
Related: #93074
Releases: master
Change-Id: If419d378d6e98c150b219fe441b0766d12e9b639