[WIP][FEATURE] Introduce lit-html and lit-element as client-side template engine
Add the lit-html template engine, that originated from
the polymer project and is now base for web components
together with the web component base case LitElement.

It is used to replace string-concatenation/jquery based
templating in order to adapt to modern web standards.

Notification, ContextMenu and LiveSearch are adapted as
first examples to show lit-html and lit-element rendering.
Notification is an example for a web component via  LitElement.
ContextMenu is now rendered directly to the
HTML element, LiveSearch renders to string.

Commands:

  yarn add lit-html@^1.3 lit-element@^2.4
  yarn add --dev grunt-rollup
  grunt build

  # todo: release for typo3/testing-framework
  composer config repositories.testing-framework vcs https://github.com/bnf/testing-framework.git
  composer require --dev typo3/testing-framework:^6.6.0@dev
  composer require --dev typo3/testing-framework:^6.6.0@dev \
    --no-update --working-dir=typo3/sysext/core

TODO:
 * Examples for lit-element and lit-html in RST

Resolves: #91810
Releases: master
Change-Id: If419d378d6e98c150b219fe441b0766d12e9b639