[TASK] Replace missing jquery import in SelectTreeElement
SelectTreeElement makes use of $.param() without importing
jquery via requireJS. Instead of adding the missing jquery
import, we simply migrate to the native URLSearchParams,
as it is available in all supported browsers.

Furthermore we avoid decoding the datastructureidentifier
json, as it is encoded to json server side. We do this,
because URLSearchParams does not url encode nested objects.

Resolves: #93200
Releases: master
Change-Id: I0da98092da6e09d6576bbdb59112d738086be238