[WIP][TASK] Migrate backend context menu to lit-html
Migrates backend UI component `ContextMenu` to be based on `lit-html`
and `lit-element`.

This also marks the `record` property and fetch() and show() methods as public,
as LegacyTree and SvgTree use these as API.

The ContextMenu is mirated to use shadow dom, that means no global
styles are used, just local ones.
Therefore the bootstrap classes are being removed and the new
IconElement custom component is used which allows to use full shadow
dom for the entire context menu.

TODO:
 * Include something https://www.npmjs.com/package/minify-html-literals
   as additonal minifier, as terser can't minify inline html and css
 * Fix submenus (disappear on hover)

Resolves: #93103
Releases: master
Change-Id: I227f44d4530d303b19efc68a077e449efe57b1e9