[WIP][FEATURE] Add support for custom HTML component based backend modules
As alternative to iframe based backend modules
HTML components can now be registered that are
used to render module contents.

A module consists of a JavaScript module that
provides an implementation for a custom HTML
tag.
This tag is applied by the new <typo3-backend-module-router>
when the route is opened via ModuleMenu (or more, @todo).

The <typo3-backend-module-router> tries to mimic
the former iframe by provided backwards compatibility
for the iframe `src` attribute and the load event (note: @todo).

This patch is currently based (and therefore incorporates):
https://review.typo3.org/c/Packages/TYPO3.CMS/+/61776/12

It is a squash of: @todo

To keep this change rather small, we do not adapt
all users of ContentContainer/inline javascript
that acceses top.list_frame. We rather rely on the
BC layer that typo3-backend-module-router to handle these
cases for now.

TODO:
 * push module state to browsers history api, or:
 * url-based routing for custom element modules
 * maybe drop module attribute from module router
   and just use src for routing (required backend
   api to register JS routes) – would only be used
   to route links to module element. We still need
   backend url generation for tokens (or not?).

Resolves: #?
Releases: master
Change-Id: I682e89649b597c8c74b6a0a8f198f6bcf5bbc347