[POC][WIP][FEATURE] Add support for custom HTML component based backend modules
This patch is currently based on (and therefore incorporates changes from):
https://review.typo3.org/c/Packages/TYPO3.CMS/+/61776/12

Furthermore this patch is a squash of:
https://github.com/TYPO3/TYPO3.CMS/compare/master...bnf:custom-component-modules-v3
Therefore, if you want to edit, please send a pull request to the next version:
https://github.com/bnf/TYPO3.CMS/commits/custom-component-modules-v4

To additionaly include include:
 * <typo3-backend-module> element (alternative for ModuleTemplate docheader)
 * Example adaption of the EXT:lowlevel configuration module as custom
   component module: Demonstrates the frontend rendering advantanges:
   Avoids scroll via buggy anchor links (docheader conflicts), as HTML
   doesn't need to be regenerated when expanding entries.

As alternative to iframe based backend modules,
HTML components can now be registered that are
used to implement modules.

A module consists of a JavaScript module that
provides an implementation for a custom HTML tag.
This tag is created by the new <typo3-backend-module-router>
when the route is opened via ModuleMenu or via deep link.

The <typo3-backend-module-router> tries to mimic
the former content "list_frame" iframe by providing backwards
compatibility for the iframe `src` attribute, the
contentWindow.location property, and (@todo) the load event.

The router currently uses two parameters for it's routing.
A named module attribute, and a endpoint (currently named src).
That means, currently it route's mostly by module name, no
client side URL-routing (yet – that's up to research/discussion)

To keep this change rather small, we do not adapt
all users of ContentContainer/inline javascript
that acceses top.list_frame. We rather rely on the
BC layer that typo3-backend-module-router provides,
to handle these cases for now.

TODO:
 * Add API allowed modules to push state to browsers history api
 * Maybe drop module attribute from module router
   and just use src for routing (required backend
   api to register JS routes) – would only be used
   to route links to module element. We still need
   backend url generation for tokens (or not?).
 * Reuse module elements when module data is changing
   (Note: changes whether iframe state needs to be
   pushed/replaced)
 * Link based routing (either full client side routing,
   or backend site routing, would then require a data-module
   tag for links)
 * Improve navigation container synchronization (when opening deeplinks)
   Maybe move into the router, or extend the router to
   route the navigation container as well.
   But then router rather becomes `<typo3-backend-application>`

Resolves: #?
Releases: master
Change-Id: I682e89649b597c8c74b6a0a8f198f6bcf5bbc347