[WIP][FEATURE] Add support for custom HTML component based backend modules
As alternative to iframe based backend modules
HTML components can now be registered that are
used to render module contents.

A module consists of a JavaScript module that
provides an implementation for a custom HTML
tag.
This tag is applied by the new <typo3-backend-module-router>
when the route is opened via ModuleMenu (or more, @todo).

The <typo3-backend-module-router> tries to mimic
the former iframe by provided backwards compatibility
for the iframe `src` attribute and the load event (note: @todo).

This patch is currently based (and therefore incorporates):
https://review.typo3.org/c/Packages/TYPO3.CMS/+/61776/12

It is a squash of: @todo

TODO:
 * url-based routing for custom element modules
 * push module state to browsers history api

Resolves: #?
Releases: master
Change-Id: I682e89649b597c8c74b6a0a8f198f6bcf5bbc347