[BUGFIX] Make BroadcastMessage.fromData() idempotent
BroadcastMessage.fromData() used to nest the entire data (which
includes a payload property itself) into another wrapping payload
property.
That caused a BroadcastMessage with a payload of {foo: "bar"}
to to end up with {payload: {foo: "bar"}} as payload, when
recreated using BroadcastMessage.fromData().

BroadcastMessage.fromData() is now fixed to preserve the correct
data structure and BroadcastMessage.createCustomEvent() is adapted
to retain the previous – additional – payload subproperty in order
to preseve backwards compatbility.

Also AjaxDataHandler is adapted to create local events with
BroadcastMessage.createCustomEvent(). This is possible, now that
local and remote BroadcastMessages are compatible.

Resolves: #93270
Related: #89244
Releases: master, 10.4
Change-Id: I5e7bbf7ea80f6e1f8951d5f32f957d8743230c01