[BUGFIX] Make BroadcastMessage.fromData() idempotent
BroadcastMessage.fromData() used to nest the entire data (which
includes a payload property itself) into another wrapping payload
property.
That caused a BroadcastMessage with a payload of {foo: "bar"}
to end up with {payload: {foo: "bar"}} as payload, when
recreated using BroadcastMessage.fromData().

BroadcastMessage.fromData() is now fixed to preserve the correct
data structure.

The Changelog documentation s fixed to refer to the correct
BroadcastMessage. The API has been explicitly marked as internal
and as possibly changing at any time, therefore it's not considered
breaking to fix this as a documentation bug.

Also AjaxDataHandler is adapted to create local events with
BroadcastMessage.createCustomEvent(). This is possible, now that
local and remote BroadcastMessages are compatible.
SlugService is adapted to use the correct BroadcastMessage structure,
as it used to send notifications as broadcasts prior to #95123.

Resolves: #93270
Related: #89244
Releases: master
Change-Id: I5e7bbf7ea80f6e1f8951d5f32f957d8743230c01