[POC][WIP][FEATURE] Introduce MFA in core
This is a POC on how MFA could be integrated in TYPO3. And
on how to provide a simple API for extensions to integrate
further MFA services / providers. The POC currently only
provides a MFA implementation for the backend but already
contains necessary preparation for the frontend integration.

Currently OTP (TOTP and HOTP) are added to the core as first
MFA provider. Furthermore is a special provider "recovery codes"
available as fallback. This provider can be setup as soon as one
of the remaining providers is activated.

The POC uses the term "MFA provider" instead of "MFA services"
to avoid confusion since there are already the authentication
services.

The idea in short:

There is an interface, which all MFA providers have to implement.
By implementing, they will be automatically registered and the API
can ensure all necessary methods exist and can be called. To access
the registered providers, the MfaProviderRegistry can be used.

The mentioned interface contains a lot of methods to be implemented
by the providers. This can be splitted up into  state providing ones,
e.g. isActive or isLocked  and the functional ones, e.g. activate or
update. Their exact task is explained in the corresponding PHPDoc.

There is furthermore the getManifest() method in place, returning
a MfaProviderManifestInterface, which contains information about
the provider such as title or description. These information are
used in various views.

There are two new controllers: The MfaConfigurationController which
is responsible for the backend module, where the user can configure
his providers and the MfaController which is accessible via the new
route "/auth/mfa", handling the multi-factor authentication process.
This route is called after a successful authentication with username
and password by the LoginController::checkRedirect() method. However
the route is only called if the user has activated at least one provider
and has not already passed the multi-factor authentication, which is
checked by a flag in the user session data.

Both controllers provide the general handling for the MFA requests
but forward the full PSR-7 request along with the user object to the
corresponding provider. The provider then, depending on the action,
either processes the requests and returns just the status (e.g.
successful or failure) or returns content which is embedded in the
controllers template. This is necessary because TYPO3 does not know
which fields or options are necessary for a provider to work.

For example, the OTP providers render the secret, a QR-code and an
input field in the setup view. Furthermore does the HOTP provider
contain a "Resync" functionality in the edit view. This means, the
main template (the one from the controller), only contains a form
whose content is then defined by the corresponding provider.

The same goes for the data, a provider needs to store. The POC
therefore uses a new database column called "mfa", which can be
freely used by the providers. The field contains of a JSON encoded
Array with each provider as array key. Common properties of such
provider array could be "active" or "lastUsed". Since the information
is stored in either the `be_users` or the `fe_users` table, the
context is implicit. Same goes for the user, the providers deal with.

It's important to have such generic field so providers are able to
store arbitrary data, TYPO3 does not need to know about. For the OTP
providers, this is e.g. the "secret" and the current "counter" value.
There is also already the MfaProviderPropertyManager in place, providing
easy access to these data with e.g. a getProperty() method.

The Manager can either be fetched from the user object directly with
`$user->getMfaProviderPropertyManager()` or by instantiating.

The MFA configuration module can be accessed in the user settings
(ext:setup) using the new field type "mfa". This type currently
simply renders a link to the module or an information, if no MFA
provider is available.

Furthermore does this POC implement following features:

* Lock functionality to prevent brute force attacks
* Default provider stored in user UC
* Switching between activated providers in the authentication view
* Resync of HOTP
* (Automatically) regneration of recovery codes
* Integration in configuration module

Following 3rd party packages are used for the POC:

* christian-riesen/base32
* endroid/qr-code

More information can be gathered by looking into the code, where
a lot of code comments are present.

Please also get in touch, e.g. in the #mfa-in-core Slack channel.

Resolves: #?
Releases: master
Change-Id: I4e902be624c80295c9c0c3286c90a6a680feeb5d