[WIP][TASK] Update to lit-html v2 and lit-element v3
Patch is currenlty WIP as the final v2 and v3 version are not
released yet. Therefore this is using pre-releases for now.

This WIP patch is here to a) see needed changes for
the upcoming releases and b) get prepared and to
avoid introducing new stuff that relies on features/
functionality that is about to be removed/deprecated.

@babel/plugin-proposal-nullish-coalescing-operator,
@babel/plugin-proposal-optional-chaining and
@rollup/plugin-babel are added because javascript
null coalescing operator and optional chaining (`?.`)
are used by the next lit version and these features
are not supported by the selenium-chrome version used for
bamboo testing right now.

Migration to the new `lit` entry point module is left
for a separate patch.
See https://github.com/Polymer/lit-html/tree/lit-next/packages/lit
for more inforamtion:
  > lit: The main module exports the core pieces needed
  > for component development: LitElement, html, css, and the most

Commands used:

  rm -rf typo3/sysext/core/Resources/Public/JavaScript/Contrib/lit-{element,html}/
  yarn add lit-html@next-major lit-element@next-major
  yarn add --dev @rollup/plugin-replace @rollup/plugin-babel \
    @babel/plugin-proposal-nullish-coalescing-operator \
    @babel/plugin-proposal-optional-chaining
  yarn build
  git add typo3/sysext/core/Resources/Public/JavaScript/Contrib/{@lit,lit-element,lit-html}/

  # todo use upstream 6.x.0 release
  composer config repositories.testing-framework vcs https://github.com/bnf/testing-framework.git
  composer require --dev typo3/testing-framework:^6.7.0@dev
  composer require --dev typo3/testing-framework:^6.7.0@dev \
    --no-update --working-dir=typo3/sysext/core

Resolves: #
Releases: master
Change-Id: I9b659d851e6ad9dc3cc649bd40aab886b86fb0f8