[TASK] Add redirect to route after login
The login controller now allows to set two new GET parameters
"redirect" (pointing to a valid route identifier) and "redirectParams"
(a list of query arguments which are rawurlencode'd).

This way, the login controller can redirect to another route
instead of dealing with a hard-coded "redirectURL", which can
be phased out in the future.

This is a first step towards allowing a defined redirect route
for all other places in the TYPO3 Backend, where currently
"redirectURL" is used. Also "linkThisScript" should become obsolete
after this change.

This change also removes the need to store a first click in
the install tool when loading the distributions within the "uc".

Resolves: #93674
Releases: master
Change-Id: Ic00113e528d00ab97215d750e1826d1ae8467a32