[POC][WIP][FEATURE] Add support for custom HTML component based backend modules
This patch is currently based on (and therefore incorporates changes from):
https://review.typo3.org/c/Packages/TYPO3.CMS/+/61776/14

As alternative to iframe based backend modules,
HTML components can now be registered that are
used to implement modules.

A module consists of a JavaScript module that
provides an implementation for a custom HTML tag.
This tag is created by the new <typo3-backend-module-router>
when the route is opened via ModuleMenu or via deep link.

The router currently uses two parameters for it's routing.
A named module attribute, and a endpoint (currently named src).
That means, currently it route's mostly by module name, no
client side URL-routing (yet – that's up to research/discussion)

State is managed via the traditional iframe (instead of history.pushState).
That may looks very strange, but we cannot combine implicit iframe
history state changes with history.pushState – iframes will take
precedence. Therefore we have to stick with iframe-state handling.
The advantage is: All "tranditional" modules will preserve the
state handling as before.

To keep this change rather small, we do not adapt
all users of ContentContainer/inline javascript
that acceses top.list_frame. We rather rely on iframe
events handlers in order to synchronize the router.

TODO:
 * Rename src-attribute to endpoint (or similar). src
   was needed in a prior patchset, in order to provide
   iframe compat, but iframe compat is now enabled
   by not hiding the iframe in
 * Enhance API to allow modules to push and/or replace state
 * Link based routing (either full client side routing,
   or backend site routing, would then require a data-module
   tag for links)
 * Improve navigation container synchronization (when opening deeplinks)
   Maybe move into the router, or extend the router to
   route the navigation container as well.

Resolves: #?
Releases: master
Change-Id: I682e89649b597c8c74b6a0a8f198f6bcf5bbc347