[WIP][FEATURE] Support Web Component based backend modules
As an alternative to iframe based backend modules,
Web (HTML) components can now be registered as
providers for the backend module view.

Such module consists of a JavaScript module that
provides an implementation for a custom HTML tag.

This tag is created by the new <typo3-backend-module-router>
when the route is opened via ModuleMenu or via a deep link.

The router y uses two parameters for it's routing.
A named module attribute, and an endpoint.
The module is used to route to a the backend module component
while the endpoint may be used by the module to parse/fetch the
current state of the module.

The router uses named slots to route between the module elements,
that means only one if it's childNodes will be visible.

<typo3-backend-module-router module="typo3-configuration-module">
  <typo3-iframe-module slot="iframe"></.…>
  <typo3-configuration-module slot="…"></…>
</typo3-backend-module-router>

State is managed via the traditional iframe (that is unlike other
modern Single Page Applications which use history.pushState).
That may looks very strange, but we cannot combine implicit iframe
history state changes with history.pushState – iframes will take
precedence. Therefore we have to stick with iframe-based state handling.
The advantage is: All "tranditional" modules will preserve the
state handling as before. No breaking/behavioral changes for traditional
modules.

As a drive-by change a new feature is added which reflects
the current module endpoint as a sharable URL in the browsers
addressbar.

TODO:
 * Link based routing (either full client side routing,
   or backend site routing, would then require a data-module
   tag for links)
 * Improve navigation container synchronization (when opening deeplinks)
   Maybe move into the router, or extend the router to
   route the navigation container as well.
 * RST

Resolves: #93988
Releases: master
Change-Id: I682e89649b597c8c74b6a0a8f198f6bcf5bbc347