[WIP][FEATURE] Support Web Component based backend modules
Web Components are now enabled to be used as an alternative
to iframe based backend modules.
Such module consists of a JavaScript module that
provides an implementation for a custom HTML tag.

As a drive-by change the browser address bar is synchronized
to reflect the current module (state) via a sharable URL.

Modules are managed by a new <typo3-backend-module-router>
component that dispatches updates from ModuleMenu or deep links
to the individual module implementations.

The router uses two parameters for it's routing: "module" and "endpoint"
 * module – Module name as defined in ext_tables.php
 * endpoint – (json) api to be used by the component (=module URL)
The module is used to route to the relative backend module component
while the endpoint may be used by the module to parse/fetch the
current state of the module.

Routing
=======
Routing uses named slots to route between the module elements,
that means only one if it's childNodes will be visible at a time.

<typo3-backend-module-router module="typo3-configuration-module">
  #shadow-root
    <slot name="typo3-configuration-module"></slot>
  <typo3-iframe-module slot="typo3-iframe-module"></…>
  <typo3-configuration-module slot="typo3-configuration-module"></…>
</typo3-backend-module-router>

History Management
==================
Browser history state is managed via the traditional iframe
history context (that is unlike other modern Single Page Applications
which use history.pushState and history.replaceState).
That may look very strange, but there are technical limitations
that prevent a combination of iframe history updates and
history.pushState. (that is: state by history.pushState is skipped
by iframe history handling).

The advantage of this compromise is: All "tranditional" modules will
preserve with their state handling as before. No breaking/behavioral
changes for iframe based modules.

TODO/Out of scope
============
 * Link based routing interception (via data-module tag for links)
 * Improve navigation container synchronization (when opening deeplinks)
   Maybe move into the router, or extend the router to route the
   navigation container as well.
 * Connect typo3-module-load(ed) events and the Loader(indicator) module
 * RST

Resolves: #93988
Releases: master
Change-Id: I682e89649b597c8c74b6a0a8f198f6bcf5bbc347