[TASK] Decouple PageTree from NavigationComponent
In order to use the non-iframe navigation component properly,
the current ties to "PageTree" (which was de-facto the only possible
navigation component to be used until TYPO3 v11 due to heavy
coupling) are removed.

The TreeInterface, which served as a wrapper for the PageTree
Navigation Component is removed, and replaced by the NavigationComponent
interface in TypeScript.

In addition, loading and initializing (and keeping state) of the
used navigation component now solely relies in the NavigationContainer.

This in turn, finally allows to e.g. properly re-load the FileStorageTree
and PageTree separately (see BackendUtility change) while
incorporating some ES6 nullable functionality.

In addition, the NavigationContainer is now jQuery-free.

Resolves: #93672
Releases: master
Change-Id: Ia16a3e789459225b21fa12efd1db6f3b4500cfef