[TASK] Streamline backend UriBuilder dependency injection
The Router dependency should not be optional and rather always be
injected, and in reality it always was, as the nullable fallback
to GeneralUtility::makeInstance was only used in unit tests.

The UriBuilder factory is now provided via the backend
service provider. That is because it became an ondemand
dependency for the backend request handler in #93674.
We now properly inject it via constructor which makes it
a lowlevel service (a service that needs to defined via
service provider to be available when TYPO3 has not yet
been installed).

Releases: master
Resolves: #94083
Change-Id: I75b39cdae8c41d1ce38860c50713d749bbe8e348