[WIP][TASK] Ensure module URLs are auto-framed by TYPO3 main controller
Or in other words:
Render module menu as links instead of buttons.

If a module is opened in a new tab it is now framed into
the backend frame. This is detected using the
Sec-Fetch-Dest which sends a "document" value
if opened as top frame. This allows to differentiate
the request from usage inside an iframe or from xhr/fetch()
requests.

Long term goal is to improve accessibility by using real
links in module menu (as we had in TYPO3 v7) and to use
module links actually every, actually everywhere, which means
the user can decide whether to open in current or a new tab.

Furthermore this intends to replace
  <a href="#" data-dispatch-action="TYPO3.ModuleMenu.showModule"
              data-dispatch-args-list="…">
with real hyperlinks…

@todo: support framing non-modules like edit-controller,
requires knowledge whether a route is a top-level route
(main,login,logout…) or not..

Resolves: #
Releases: master
Change-Id: Iaad8fe62644ec6c9cb1aaa4c92ce2e8e3eeab7bd