[WIP][TASK] Make module requests Sec-Fetch-Dest aware
This allows module URLs to automatically framed by the TYPO3
main controller whenever they are directly requested via the
browser address bar (Header: Sec-Fetch-Dest=document). See:
https://w3c.github.io/webappsec-fetch-metadata/#sec-fetch-dest-header

Note: This header already supported by Chrome and Firefox nightly,
Safari does not deliver this header yet.
Non supporting browsers will gracefully fallback to showing
the module contents without the TYPO3 main frame being added.
This fallback behaviour is only triggered when a link is
opened in a new tab. Regular clicks are intercepted and
directly dispatched via module-router/iframe, as before.

This allows to render the module menu with anchor tags
instead of buttons which lets users decide how and
where a module should be opened.

Long term goal is to improve accessibility by using real
links not only in module menu (as we had in TYPO3 v7), but
actually everywhere (for example in the databse record list),
in order for the user to decide whether to open a link in
the current or a new tab.

Now, when that is done a global click event handler can be added
that allows to replace
  <a href="#" data-dispatch-action="TYPO3.ModuleMenu.showModule"
              data-dispatch-args-list="…">
with real hyperlinks, but that is out of scope for this change.

TODO
====

 * Support framing of non-modules routes like edit-controller.
   This requires knowledge whether a route is a top-level route
   (main,login,logout…) or not..
   Or maybe we should add module-specific routes for the
   edit-controller?
   That brings another advantage, as sharable links would include
   the selected module menu entry.

Resolves: #94084
Releases: master
Change-Id: Iaad8fe62644ec6c9cb1aaa4c92ce2e8e3eeab7bd